.claude/skills/perfil-rrhh-pdf-word-pdf/requirements.txt
pdfplumber>=0.11.0
python-docx>=1.1.0
docx2pdf>=0.1.8
mammoth>=1.6.0
weasyprint>=62.0
